requirements.txt
Django==5.2.7
supabase==2.23.0
psycopg2-binary==2.9.11
python-decouple==3.8
dj-database-url==2.3.0
Pillow==12.0.0
